+++ frontend/controllers/CompanyController.php
@@ -10,18 +10,38 @@ use yii\web\NotFoundHttpException;
 use yii\filters\VerbFilter;
 use yii\helpers\Url;
 use yii\web\UploadedFile;
+use yii\filters\AccessControl;
+use yii\web\ForbiddenHttpException;
 
 /**
  * CompanyController implements the CRUD actions for Company model.
  */
 class CompanyController extends Controller
 {
+   // public $layout='initial';
     /**
      * @inheritdoc
      */
+
     public function behaviors()
     {
         return [
+            'access' => [
+                'class' => AccessControl::className(),
+                'only' => ['view', 'index','create','update','delete'],
+                'rules' => [
+                    [
+                        'actions' => ['view','index'],
+                        'allow' => true,
+                        'roles' => ['?'],
+                    ],
+                    [
+                        'actions' => ['view','index','create','update','delete'],
+                        'allow' => true,
+                        'roles' => ['@'],
+                    ],
+                ],
+            ],
             'verbs' => [
                 'class' => VerbFilter::className(),
                 'actions' => [
@@ -70,37 +90,44 @@ class CompanyController extends Controller
         date_default_timezone_set('Asia/Kolkata');
         $current_date = date("Y-m-d h:i:sa");
         $model->company_created = $current_date;
-        
-        if ($model->load(Yii::$app->request->post())) {
-            $imageName= $model->company_name;
 
-            $model->file = UploadedFile::getInstance($model,'file');
-            if($model->file->extension=='gif'||$model->file->extension=='jpg'||$model->file->extension=='png')
-            {
-                $model->file->saveAs('uploads/'.$imageName.'.'.$model->file->extension);
+        if(Yii::$app->user->can('create_company'))//Authenticate whether user have right to create company
+        {
+            if ($model->load(Yii::$app->request->post())) {
+                $imageName= $model->company_name;
 
-                // save the path in DB.
+                $model->file = UploadedFile::getInstance($model,'file');
+                if($model->file->extension=='gif'||$model->file->extension=='jpg'||$model->file->extension=='png')
+                {
+                    $model->file->saveAs('uploads/'.$imageName.'.'.$model->file->extension);
 
-                $model->company_profile = 'uploads/'.$imageName.'.'.$model->file->extension;
+                    // save the path in DB.
+
+                    $model->company_profile = 'uploads/'.$imageName.'.'.$model->file->extension;
+
+                    
+                    $model->save();
+                    return $this->redirect(['view', 'id' => $model->company_id]);
+                }
+                else{
+                    return $this->render('create', [
+                    'model' => $model,'msg'=>'Please upload image file only',
+                ]);
+                }
 
                 
-                $model->save();
-                return $this->redirect(['view', 'id' => $model->company_id]);
-            }
-            else{
+            }  
+            else 
+            {
                 return $this->render('create', [
-                'model' => $model,'msg'=>'Please upload image file only',
-            ]);
+                    'model' => $model,
+                ]);
             }
-
-            
-        }  
-        else 
-        {
-            return $this->render('create', [
-                'model' => $model,
-            ]);
         }
+        else{
+            throw new ForbiddenHttpException('You are not permitted to do this action');
+        }
+
     }
 
     
@@ -114,42 +141,47 @@ class CompanyController extends Controller
     public function actionUpdate($id)
     {
         $model = $this->findModel($id);
-        
-        if ($model->load(Yii::$app->request->post())) 
+        if(Yii::$app->user->can('update_company'))//Authenticate whether user have right to update company
         {
-            
-            $imageName= $model->company_name;
-
-            $model->file = UploadedFile::getInstance($model,'file');
-            if($model->file!='')
+            if ($model->load(Yii::$app->request->post())) 
             {
+                
+                $imageName= $model->company_name;
 
-             if(($model->file->extension=='gif'||$model->file->extension=='jpg'||$model->file->extension=='png'))
+                $model->file = UploadedFile::getInstance($model,'file');
+                if($model->file!='')
                 {
 
-                    $model->file->saveAs('uploads/'.$imageName.'.'.$model->file->extension);
+                 if(($model->file->extension=='gif'||$model->file->extension=='jpg'||$model->file->extension=='png'))
+                    {
 
-                    // save the path in DB.
+                        $model->file->saveAs('uploads/'.$imageName.'.'.$model->file->extension);
 
-                    $model->company_profile = 'uploads/'.$imageName.'.'.$model->file->extension;
-                    
+                        // save the path in DB.
+
+                        $model->company_profile = 'uploads/'.$imageName.'.'.$model->file->extension;
+                        
+                        $model->save();
+                        return $this->redirect(['view', 'id' => $model->company_id]);
+                    }
+                    else
+                    {
+                        return $this->render('update', ['model' => $model,'msg'=>'Please upload image file only',]);
+                    } 
+                }
+                else{
                     $model->save();
                     return $this->redirect(['view', 'id' => $model->company_id]);
                 }
-                else
-                {
-                    return $this->render('update', ['model' => $model,'msg'=>'Please upload image file only',]);
-                } 
             }
-            else{
-                $model->save();
-                return $this->redirect(['view', 'id' => $model->company_id]);
+            else {
+                return $this->render('update', [
+                    'model' => $model,
+                ]);
             }
         }
-        else {
-            return $this->render('update', [
-                'model' => $model,
-            ]);
+        else{
+            throw new ForbiddenHttpException('You are not permitted to do this action');
         }
     }
 
@@ -161,9 +193,15 @@ class CompanyController extends Controller
      */
     public function actionDelete($id)
     {
-        $this->findModel($id)->delete();
+        if(Yii::$app->user->can('delete_company'))//Authenticate whether user have right to delete company
+        {
+            $this->findModel($id)->delete();
 
-        return $this->redirect(['index']);
+            return $this->redirect(['index']);
+        }
+        else{
+            throw new ForbiddenHttpException('You are not permitted to do this action');
+        }
     }
 
     /**
