+++ frontend/controllers/DepartmentController.php
@@ -9,12 +9,30 @@ use yii\web\NotFoundHttpException;
 use yii\filters\VerbFilter;
 use frontend\models\Company;
 use frontend\models\Branch;
+use yii\filters\AccessControl;
+use yii\web\ForbiddenHttpException;
 
 class DepartmentController extends \yii\web\Controller
 {
 	public function behaviors()
     {
         return [
+            'access' => [
+                'class' => AccessControl::className(),
+                'only' => ['view', 'index','create','update','delete'],
+                'rules' => [
+                    [
+                        'actions' => ['view','index'],
+                        'allow' => true,
+                        'roles' => ['?'],
+                    ],//guest user can only view department
+                    [
+                        'actions' => ['view','index','create','update','delete'],
+                        'allow' => true,
+                        'roles' => ['@'],
+                    ],// authenticated user can only create update delete department according to authmanager
+                ],
+            ],
             'verbs' => [
                 'class' => VerbFilter::className(),
                 'actions' => [
@@ -43,20 +61,26 @@ class DepartmentController extends \yii\web\Controller
      */
     public function actionCreate()
     {
-         $model = new Department();
-         $company = Company::find()->orderBy('company_name')->all(); 
-         $branch = Branch::find()->orderBy('branch_name')->all(); 
-        date_default_timezone_set('Asia/Kolkata');
-        $current_date = date("Y-m-d h:i:sa");
-        $model->department_created = $current_date;
-        if ($model->load(Yii::$app->request->post()) && $model->save()) {
-            return $this->redirect(['view', 'id' => $model->department_id]);
-        } else {
-            return $this->render('create', [
-                'model' => $model,'company'=>$company,'branch'=>$branch,
-            ]);
-        }
+        if(Yii::$app->user->can('create_department'))//Authenticate whether user have right to create department
+        {
+            $model = new Department();
+            $company = Company::find()->orderBy('company_name')->all(); 
+            $branch = Branch::find()->orderBy('branch_name')->all(); 
+            date_default_timezone_set('Asia/Kolkata');
+            $current_date = date("Y-m-d h:i:sa");
+            $model->department_created = $current_date;
 
+            if ($model->load(Yii::$app->request->post()) && $model->save()) {
+                return $this->redirect(['view', 'id' => $model->department_id]);
+            } else {
+                return $this->render('create', [
+                    'model' => $model,'company'=>$company,'branch'=>$branch,
+                ]);
+            }
+        }
+        else{
+            throw new ForbiddenHttpException('You are not permitted to do this action');
+        }
         
     }
 
@@ -66,15 +90,21 @@ class DepartmentController extends \yii\web\Controller
      */
     public function actionUpdate($id)
     {
-        $model = $this->findModel($id);
-        $company = Company::find()->orderBy('company_name')->all(); 
-        $branch = Branch::find()->orderBy('branch_name')->all(); 
-        if ($model->load(Yii::$app->request->post()) && $model->save()) {
-            return $this->redirect(['view', 'id' => $model->department_id]);
-        } else {
-            return $this->render('update', [
-                'model' => $model,'company'=>$company,'branch'=>$branch,
-            ]);
+        if(Yii::$app->user->can('update_department'))//Authenticate whether user have right to update department
+        {
+            $model = $this->findModel($id);
+            $company = Company::find()->orderBy('company_name')->all(); 
+            $branch = Branch::find()->orderBy('branch_name')->all(); 
+            if ($model->load(Yii::$app->request->post()) && $model->save()) {
+                return $this->redirect(['view', 'id' => $model->department_id]);
+            } else {
+                return $this->render('update', [
+                    'model' => $model,'company'=>$company,'branch'=>$branch,
+                ]);
+            }
+        }
+        else{
+            throw new ForbiddenHttpException('You are not permitted to do this action');
         }
     }
 
@@ -107,9 +137,15 @@ class DepartmentController extends \yii\web\Controller
      */
     public function actionDelete($id)
     {
-        $this->findModel($id)->delete();
+        if(Yii::$app->user->can('delete_department'))//Authenticate whether user have right to delete department
+        {
+            $this->findModel($id)->delete();
 
-        return $this->redirect(['index']);
+            return $this->redirect(['index']);
+        }
+        else{
+            throw new ForbiddenHttpException('You are not permitted to do this action');
+        }
     }
 
 }
